Split methods, and hide assertions behind test feature flag
